Function supports g.analyse. Not intended for direct use by user.

Generates recording specific analyses and fills corresponding output matrix, g.analyse.

Usage

g.analyse.perfile(I, C, metrics_nav,
                 AveAccAve24hr, doquan, doiglevels, tooshort,
                 params_247, params_cleaning, params_general,
                 output_avday, output_perday,
                 dataqual_summary, file_summary)

Arguments

I

output g.inspectfile

C

output g.calibrate

metrics_nav

List with three objects to help navigate the acceleration metrics

AveAccAve24hr

Average acceleration in an average 24 hour cycle

doquan

Boolean whether quantile analysis should be done

doiglevels

Boolean to indicate whether iglevels should be calculated

tooshort

0 (file not too short) or 1 (file too short)

params_247

see GGIR

params_cleaning

see GGIR

params_general

see GGIR

output_avday

Output from g.analyse.avday

output_perday

Output from g.analyse.perday

dataqual_summary

Data.frame with data quality summary indicators produced in g.analyse

Value

filesummary

summary for the file that was analysed

daysummary

Summary per day for the file that was analysed

Author

Vincent T van Hees <v.vanhees@accelting.com>