Generate report from milestone data produced by g.part2

Creates report from milestone data produced by g.part2. Not intended for direct use by package user

Usage

g.report.part2(metadatadir = c(), f0 = c(), f1 = c(), maxdur = 0,
  selectdaysfile = c(), store.long = FALSE, do.part2.pdf = TRUE, 
  sep_reports = ",", verbose = TRUE)

Arguments

metadatadir

Directory that holds a folder 'meta' and inside this a folder 'basic' which contains the milestone data produced by g.part1. The folderstructure is normally created by g.part1 and GGIR will recognise what the value of metadatadir is.

f0

File index to start with (default = 1). Index refers to the filenames sorted in alphabetical order

f1

File index to finish with (defaults to number of files available, i.e., f1 = 0)

maxdur

see g.part2

selectdaysfile

see g.part2

store.long

Booelean to indicate whether output should stored in long format in addition to default wide format. Automatically turned to TRUE if using day segmentation with qwindow.

do.part2.pdf

Boolean, see g.part2

sep_reports

Character (default = ","). Value used as sep argument in write.csv for writing csv files containing the GGIR reports.

verbose

See details in GGIR.

Value

Function does not produce data, but only writes reports in csv format and visual reports in pdf format

Author

Vincent T van Hees <v.vanhees@accelting.com>